extract html table creation into separate file
