+++ electron/run.js
@@ -1,9 +1,9 @@
 'use strict'
 
 const { load, setEntResults } = require('./graph')
+const makeTableFromObjectList = require('./lib/html-table')
 const prettyMs = require('pretty-ms')
 const fs = require('fs')
-const toPairs = require('lodash/toPairs')
 
 function run() {
 	// get the file
@@ -51,54 +51,6 @@ function submitJob({ socket = global.socket, pipeline, filepath, data }) {
 	})
 }
 
-function makeTableFromObject(data) {
-	let table = document.createElement('table')
-
-	let thead = document.createElement('thead')
-	let tr = document.createElement('tr')
-
-	let first = data[0]
-	if (!first) {
-		return table
-	}
-
-	let keys = Object.keys(first).filter(key => !key.startsWith('__'))
-
-	for (let key of keys) {
-		let th = document.createElement('th')
-		th.innerHTML = key
-		tr.appendChild(th)
-	}
-
-	thead.appendChild(tr)
-	table.appendChild(thead)
-
-	let tbody = document.createElement('tbody')
-	for (let distribution of data) {
-		let tr = document.createElement('tr')
-
-		if (distribution.__highlight) {
-			tr.classList.add('highlight')
-		}
-
-		let values = toPairs(distribution)
-			.filter(([key]) => !key.startsWith('__'))
-			.map(([_, value]) => value)
-
-		for (let value of values) {
-			let td = document.createElement('td')
-			td.innerHTML = value
-			tr.appendChild(td)
-		}
-
-		tbody.appendChild(tr)
-	}
-
-	table.appendChild(tbody)
-
-	return table
-}
-
 function onData(phase, data) {
 	console.info([phase, data])
 	if (phase.startsWith('newick-json:')) {
@@ -125,11 +77,11 @@ function onData(phase, data) {
 
 		document
 			.querySelector('#distributions')
-			.appendChild(makeTableFromObject(data.distributions))
+			.appendChild(makeTableFromObjectList(data.distributions))
 
 		document
 			.querySelector('#probabilities')
-			.appendChild(makeTableFromObject(data.probabilities))
+			.appendChild(makeTableFromObjectList(data.probabilities))
 
 		data.results = data.results.map(item => {
 			if (item.Probability < 0.05) {
@@ -139,7 +91,7 @@ function onData(phase, data) {
 		})
 		document
 			.querySelector('#results')
-			.appendChild(makeTableFromObject(data.results))
+			.appendChild(makeTableFromObjectList(data.results))
 	} else if (phase === 'nonmonophyletic-sequences') {
 		setEntResults(data)
 	} else {
